Refactor Resonite live-send execution helpers
